UPDATE : not functional
Ok so there is something weird going on where the import statement to get tensorflow.python.data Dataset is not working and I have no idea what is going on and or why it is causeing so many problems. If somone would please figure out what is going on that would be greatly appreciated.